Move ia2_get_pkey into libia2
We need to use ia2_get_pkey from outside the allocator library, so this
change moves that function into libia2 where it can be shared for both.
