+++ libia2/ia2.c
@@ -5,6 +5,74 @@
 
 #include "ia2.h"
 
+#ifdef LIBIA2_INSECURE
+size_t ia2_get_pkey() { return 0; }
+#else
+size_t ia2_get_pkey() {
+  uint32_t pkru;
+  __asm__("rdpkru" : "=a"(pkru) : "a"(0), "d"(0), "c"(0));
+  switch (pkru) {
+  case 0xFFFFFFFC: {
+    return 0;
+  }
+  case 0xFFFFFFF0: {
+    return 1;
+  }
+  case 0xFFFFFFCC: {
+    return 2;
+  }
+  case 0xFFFFFF3C: {
+    return 3;
+  }
+  case 0xFFFFFCFC: {
+    return 4;
+  }
+  case 0xFFFFF3FC: {
+    return 5;
+  }
+  case 0xFFFFCFFC: {
+    return 6;
+  }
+  case 0xFFFF3FFC: {
+    return 7;
+  }
+  case 0xFFFCFFFC: {
+    return 8;
+  }
+  case 0xFFF3FFFC: {
+    return 9;
+  }
+  case 0xFFCFFFFC: {
+    return 10;
+  }
+  case 0xFF3FFFFC: {
+    return 11;
+  }
+  case 0xFCFFFFFC: {
+    return 12;
+  }
+  case 0xF3FFFFFC: {
+    return 13;
+  }
+  case 0xCFFFFFFC: {
+    return 14;
+  }
+  case 0x3FFFFFFC: {
+    return 15;
+  }
+  // TODO: We currently treat any unexpected PKRU value as pkey 0 (the shared
+  // heap) for simplicity since glibc(?) initializes the PKRU to 0x55555554
+  // (usually). We don't set the PKRU until the first compartment transition, so
+  // let's default to using the shared heap before our first wrpkru. When we
+  // initialize the PKRU properly (see issue #95) we should probably abort when
+  // we see unexpected PKRU values.
+  default: {
+    return 0;
+  }
+  }
+}
+#endif // LIBIA2_INSECURE
+
 static const char *shared_sections[][2] = {
     {"__start_ia2_shared_data", "__stop_ia2_shared_data"},
 };
