+++ scripts/partition_alloc/partition_alloc.diff
@@ -7,7 +7,7 @@ index 6434ee54..c27030dd 100644
  #endif  // BUILDFLAG(STARSCAN)
  
 +#include <sys/mman.h>
-+#include <ia2_get_pkey.h>
++#include <ia2.h>
 +
  namespace partition_alloc::internal {
